requirements.txt
streamlit
pandas 
pymysql
plotly
python-dotenv